Workaround for Flash Player 13 changes
FP13 in non-Chrome browsers (PPAPI Flash seems unaffected) ignores the
contentType parameter in the URLRequest object when performing a
navigateToURL request, resulting in an inability to upload the recording
as the server will try to interpret the multipart data as urlencoded vars.

By changing to URLLoader.load(), we can get around this restriction, but
the page does not load in the browser window - so a new flashvar has been
introduced - return.  The return flashvar takes a URL, which the user will
be redirected to upon upload.  Setting the "return" flashvar to "reset"
results in the recorder being reset for a new recording rather than the
browser being redirected to a URL; omitting it altogether will also reset
the recorder if you are NOT using a JS callback.